refactor(application): extract initApplicationService and move templates
