+++ src/application/impl/templates/resume_pool.md
