+++ src/application/impl/templates/standard_questions.md
