+++ src/application/impl/templates/workspace-claude.md
